% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogues_ratio.R
\name{calculate_paralogues_ratio}
\alias{calculate_paralogues_ratio}
\title{Calculate ratio for Paralogues}
\usage{
calculate_paralogues_ratio(paralogues, input_genes)
}
\arguments{
\item{paralogues}{A df of paralogue annotations for all protein coding genes}

\item{input_genes}{A vector of input genes}
}
\value{
A dataframe with HGNC IDs and paralogue score
}
\description{
Calculate ratio for Paralogues
}
