% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10PpiRatio.R
\name{calculate_ppi_ratio}
\alias{calculate_ppi_ratio}
\title{Calculate ratio for PPI}
\usage{
calculate_ppi_ratio(ppi, input_genes)
}
\arguments{
\item{ppi}{A df of ppi annotations for all protein coding genes}

\item{input_genes}{A df of input genes}
}
\value{
A dataframe with HGNC IDs and PPI scores
}
\description{
Calculate ratio for PPI
}
