% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step11ProteinComplexRatio.R
\name{calculate_protein_complex_ratio}
\alias{calculate_protein_complex_ratio}
\title{Calculate ratio for Protein Complex}
\usage{
calculate_protein_complex_ratio(complexportal, input_genes)
}
\arguments{
\item{complexportal}{A df of protein complex annotations of input genes}

\item{input_genes}{A vector of input genes}
}
\value{
A dataframe with HGNC IDs and protein complex score
}
\description{
Calculate ratio for Protein Complex
}
