% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_families_ratio.R
\name{calculate_protein_families_ratio}
\alias{calculate_protein_families_ratio}
\title{Calculate ratio for Protein Families}
\usage{
calculate_protein_families_ratio(panther, input_genes)
}
\arguments{
\item{panther}{A df of protein family annotations of input genes}

\item{input_genes}{A vector of input genes}
}
\value{
A dataframe with HGNC IDs and protein family score
}
\description{
Calculate ratio for Protein Families
}
