% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_check.R
\name{check_paralogues}
\alias{check_paralogues}
\title{Check for Paralogues}
\usage{
check_paralogues(paralogues, input_genes)
}
\arguments{
\item{paralogues}{A df of paralogue annotations for all protein coding genes}

\item{input_genes}{A vector of input genes}
}
\value{
A dataframe with HGNC IDs and paralogues for manual check
}
\description{
Check for Paralogues
}
