% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualCheck.R
\name{check_pathways}
\alias{check_pathways}
\title{Check Pathways}
\usage{
check_pathways(protein_coding_genes, input_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}
}
\value{
A df of input genes pathway data
}
\description{
Check Pathways
}
