% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualCheck.R
\name{check_ppi}
\alias{check_ppi}
\title{Check for PPI}
\usage{
check_ppi(ppi, input_genes)
}
\arguments{
\item{ppi}{A df of ppi annotations for all protein coding genes}

\item{input_genes}{A df of input genes}
}
\value{
A dataframe with HGNC IDs and PPI scores for input genes
}
\description{
Check for PPI
}
