% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_check.R
\name{check_protein_families}
\alias{check_protein_families}
\title{Check protein families}
\usage{
check_protein_families(
  protein_coding_genes,
  uniprot_input_gene_symbol_results_cleaned
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{uniprot_input_gene_symbol_results_cleaned}{The df of uniprot results of input genes.}
}
\value{
A dataframe with Protein Family data from Uniprot for the input genes.
}
\description{
Check protein families
}
