% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2Paralogs.R
\name{fetch_paralogs}
\alias{fetch_paralogs}
\title{Fetch Paralogs Data}
\usage{
fetch_paralogs(
  protein_coding_genes,
  chunk_size = 100,
  save_raw = FALSE,
  save_path = NULL
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{chunk_size}{Integer for the size of each chunk to query from Ensembl.}

\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data.}
}
\value{
A dataframe with paralogs data from Ensembl.
}
\description{
This function fetches paralogs data from Ensembl for all protein coding genes.
}
