% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogues.R
\name{fetch_paralogues}
\alias{fetch_paralogues}
\title{Fetch Paralogues Data}
\usage{
fetch_paralogues(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}
}
\value{
A dataframe with paralogues data from Ensembl.
}
\description{
This function fetches paralogues data from Ensembl for all protein coding genes.
}
