% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{fetch_pathways}
\alias{fetch_pathways}
\title{Fetch Pathway Data}
\usage{
fetch_pathways(protein_coding_genes, input_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}
}
\value{
A list of two dataframes with pathway data from Reactome 1. input genes pathway data, 2. All protein coding genes pathway data
}
\description{
This function fetches pathway data from Reactome for all protein coding genes and input genes.
}
