% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04Ppi.R
\name{fetch_ppi}
\alias{fetch_ppi}
\title{Fetch PPI Data}
\usage{
fetch_ppi(protein_coding_genes, save_raw = FALSE, save_path = NULL)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data.}
}
\value{
A dataframe with data from Database.
}
\description{
This function fetches PPI data from STRING for all protein coding genes.
}
