% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ProteinFamilies.R
\name{fetch_protein_families}
\alias{fetch_protein_families}
\title{Fetch Protein Family Data from Uniprot}
\usage{
fetch_protein_families(
  protein_coding_genes,
  uniprot_input_gene_symbol_results_cleaned,
  save_raw = FALSE,
  save_path = NULL
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{uniprot_input_gene_symbol_results_cleaned}{The df of uniprot results of input genes.}

\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data.}
}
\value{
A dataframe with Protein Family data from Uniprot for the input genes.
}
\description{
This function fetches protein family data from Uniprot for the input genes.
}
