% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5Uniprot.R
\name{fetch_uniprot}
\alias{fetch_uniprot}
\title{Fetch Uniprot Data}
\usage{
fetch_uniprot(
  protein_coding_genes,
  input_genes,
  save_raw = FALSE,
  save_path = NULL
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}

\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data.}
}
\value{
A dataframe with Uniprot data of the input genes.
}
\description{
This function fetches protein data from Uniprot for all protein coding genes_______?.
}
