% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{fivep_distribution_plot}
\alias{fivep_distribution_plot}
\title{Function for plotting distribution of fiveP scores}
\usage{
fivep_distribution_plot(dis_plots, bins = 20)
}
\arguments{
\item{bins}{Number of bins for histogram}

\item{dis_plot}{A list of dfs of fiveP results to plot}
}
\value{
A ggplot2 plot of fiveP distribution scores
}
\description{
Function for plotting distribution of fiveP scores
}
