% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{fivep_distribution_plot_zoomedin}
\alias{fivep_distribution_plot_zoomedin}
\title{Function for plotting distribution of fiveP scores with zoomed-in y-axis}
\usage{
fivep_distribution_plot_zoomedin(dis_plots, bins = 20, ymax = 2000)
}
\arguments{
\item{bins}{Number of bins for histogram}

\item{ymax}{Maximum y-axis value for zooming in}

\item{dis_plot}{A list of dfs of fiveP results to plot}
}
\value{
A ggplot2 plot of fiveP distribution scores with zoomed-in y-axis
}
\description{
Function for plotting distribution of fiveP scores with zoomed-in y-axis
}
