% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualCheck.R
\name{gene_check}
\alias{gene_check}
\title{Main function for checks}
\usage{
gene_check(
  protein_coding_genes,
  uniprot_input_gene_symbol_results_cleaned,
  paralogues,
  ppi,
  input_genes
)
}
\arguments{
\item{protein_coding_genes}{A df of protein coding genes}

\item{uniprot_input_gene_symbol_results_cleaned}{description}

\item{paralogues}{description}

\item{ppi}{description}

\item{input_genes}{A df of input genes}
}
\value{
A dataframe with HGNC IDs and protein features for input genes
}
\description{
Main function for checks
}
