% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{plot_go_similarity}
\alias{plot_go_similarity}
\title{Function for plotting GO term similarity of input gene list vs random gene lists}
\usage{
plot_go_similarity(df, sd_col, sd_series, title = NULL)
}
\arguments{
\item{df}{A dataframe of GO similarity results to plot}

\item{sd_col}{The name of the column containing standard deviation values (as a string)}

\item{sd_series}{The name of the series to which the standard deviation applies (as a string)}

\item{title}{Optional title for the plot}
}
\value{
A ggplot2 plot of GO term similarity results
}
\description{
Function for plotting GO term similarity of input gene list vs random gene lists
}
