% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{plot_go_similarity_1v1}
\alias{plot_go_similarity_1v1}
\title{Function for plotting GO term similarity of input gene list vs one random gene lists}
\usage{
plot_go_similarity_1v1(df, title = NULL)
}
\arguments{
\item{df}{A data frame of GO similarity results with columns: threshold_label, input_list, random_list}

\item{title}{Optional title for the plot}
}
\value{
A ggplot2 plot of GO term similarity results
}
\description{
Function for plotting GO term similarity of input gene list vs one random gene lists
}
