% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{plot_go_similarity_1vmany}
\alias{plot_go_similarity_1vmany}
\title{Function for plotting GO term similarity of input gene list vs many random gene lists}
\usage{
plot_go_similarity_1vmany(df, name_of_sd_series = "", title = NULL)
}
\arguments{
\item{df}{A data frame of GO similarity results with columns: threshold_label, input_list, random_list, random_list_sd}

\item{name_of_sd_series}{Name of the series in the df for which to plot error bars}

\item{title}{Optional title for the plot}
}
\value{
A ggplot2 plot of GO term similarity results
}
\description{
Function for plotting GO term similarity of input gene list vs many random gene lists
}
