% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14WilcoxonSignedRankTest.R
\name{run_wilcox_tests}
\alias{run_wilcox_tests}
\title{Function for calculating Wilcoxon signed-rank test between two sets of GO similarity scores}
\usage{
run_wilcox_tests(vector, list)
}
\arguments{
\item{vector}{A numeric vector of GO similarity scores for the input gene list}

\item{list}{A list of numeric vectors, each representing GO similarity scores for a random gene list}
}
\value{
A data frame with Wilcoxon test results (p-values and statistics) for each comparison
}
\description{
Function for calculating Wilcoxon signed-rank test between two sets of GO similarity scores
}
