% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRawFiles.R
\name{save_raw_files}
\alias{save_raw_files}
\title{saving raw files for each P}
\usage{
save_raw_files(input_genes, save_path)
}
\arguments{
\item{input_genes}{The df of input genes.}

\item{save_path}{path to save the raw files.}
}
\value{
just a print statement.the files will be saved.
}
\description{
saving raw files for each P
}
