% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRawFiles.R
\name{save_raw_files_uniprot_only}
\alias{save_raw_files_uniprot_only}
\title{saving raw files for each P - that use uniprot}
\usage{
save_raw_files_uniprot_only(input_genes, save_path, name)
}
\arguments{
\item{input_genes}{The df of input genes.}

\item{save_path}{path to save the raw files.}

\item{name}{the name of the files to save.}
}
\value{
just a print statement.the files will be saved.
}
\description{
saving raw files for each P - that use uniprot
}
