% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1HgncGeneList.R
\name{fetch_hgnc_gene_list}
\alias{fetch_hgnc_gene_list}
\title{Fetch all protein coding genes data}
\usage{
fetch_hgnc_gene_list(save_raw = FALSE, save_path = NULL)
}
\arguments{
\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data. If NULL, it will save to "data/hgnc_gene_list.csv".}
}
\value{
A dataframe with protein coding gene data from EBI database.
}
\description{
This function fetches data from EBI to get all protein coding genes.
}
