% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{fetch_paralogs}
\alias{fetch_paralogs}
\title{Fetch Paralogs Data}
\usage{
fetch_paralogs(
  protein_coding_genes,
  chunk_size = 100,
  save_raw = FALSE,
  save_path = NULL
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}
}
\value{
A dataframe with paralogs data from Ensembl.
}
\description{
This function fetches paralogs data from Ensembl for all protein coding genes.
}
