% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3Pathways.R
\name{fetch_pathways}
\alias{fetch_pathways}
\title{Fetch Pathway Data}
\usage{
fetch_pathways(
  protein_coding_genes,
  input_genes,
  save_raw = FALSE,
  save_path = NULL
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}

\item{save_raw}{Boolean for choosing whether to save the raw data or not.}

\item{save_path}{String for the path to save the raw data.}
}
\value{
A list of two dataframes with pathway data from Reactome 1. input genes pathway data, 2. All protein coding genes pathway data
}
\description{
This function fetches pathway data from Reactome for all protein coding genes and input genes.
}
