% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{fetch_ppi}
\alias{fetch_ppi}
\title{Fetch PPI Data}
\usage{
fetch_ppi(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}
}
\value{
A dataframe with data from Database.
}
\description{
This function fetches PPI data from STRING for all protein coding genes.
}
