% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fetch_uniprot}
\alias{fetch_uniprot}
\title{Fetch Uniprot Data}
\usage{
fetch_uniprot(protein_coding_genes, input_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}
}
\value{
A dataframe with Uniprot data of the input genes.
}
\description{
This function fetches protein data from Uniprot for all protein coding genes_______?.
}
