% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13Plotting.R
\name{plot_go_similarity}
\alias{plot_go_similarity}
\title{Function for plotting GO term similarity of input gene list vs random gene lists}
\usage{
plot_go_similarity(
  list1,
  list2,
  thresholds = seq(0, 1, 0.05),
  sd_col = NULL,
  sd_series = NULL,
  title = NULL
)
}
\arguments{
\item{list1}{A numeric vector of GO similarity scores for the input gene list}

\item{list2}{A numeric vector of GO similarity scores for a random gene list}

\item{thresholds}{A numeric vector of threshold values (default is seq(0, 1, 0.05))}

\item{sd_col}{The name of the column containing standard deviation values (as a string)}

\item{sd_series}{The name of the series to which the standard deviation applies (as a string)}

\item{title}{Optional title for the plot}
}
\value{
A ggplot2 plot of GO term similarity results
}
\description{
Function for plotting GO term similarity of input gene list vs random gene lists
}
